// x-pack/solutions/observability/plugins/apm/common/service_inventory.d.ts
import type { AgentName } from '@kbn/elastic-agent-utils';
import type { SloStatus } from '@kbn/apm-types';
export type { SloStatus } from '@kbn/apm-types';
export interface ServiceListItem {
    serviceName: string;
    anomalyScore?: number;
    transactionType?: string;
    agentName?: AgentName;
    throughput?: number;
    latency?: number | null;
    transactionErrorRate?: number | null;
    environments?: string[];
    alertsCount?: number;
    overflowCount?: number | null;
    sloStatus?: SloStatus;
    sloCount?: number;
}
export declare enum ServiceInventoryFieldName {
    ServiceName = "serviceName",
    AnomalyScore = "anomalyScore",
    Environments = "environments",
    TransactionType = "transactionType",
    Throughput = "throughput",
    Latency = "latency",
    TransactionErrorRate = "transactionErrorRate",
    AlertsCount = "alertsCount",
    SloStatus = "sloStatus"
}